move cellentity
