+++ src/main/java/bomberman/Wall.java
@@ -1,5 +1,6 @@
 package bomberman;
 
+import bomberman.attributes.CellEntity;
 import bomberman.attributes.SolidEntity;
 
 public class Wall implements CellEntity, SolidEntity {
