+++ src/main/java/bomberman/attributes/CellEntity.java
@@ -1,4 +1,4 @@
-package bomberman;
+package bomberman.attributes;
 
 public interface CellEntity {
     public void interactWith(CellEntity anotherEntity);
